requirements.txt
flask==2.2.2
python-dateutil
requests
pandas==1.5.3
numpy
plotly
statsmodels